env param

+++ var/cgi-bin/uploadFile.cgi
@@ -15,6 +15,9 @@ def create_storage_directory():
 
 def save_file(storage_path, file_name, file_content):
     file_path = os.path.join(storage_path, file_name)
+    if os.path.exists(file_path):
+        print(f"409")
+        sys.exit(-1)
     try:
         with open(file_path, 'wb') as file:
             file.write(file_content)
@@ -24,12 +27,14 @@ def save_file(storage_path, file_name, file_content):
         sys.exit(-1)
 
 def main():
-    if len(sys.argv) != 3:
+    if len(sys.argv) != 1:
         print("500")
         sys.exit(-1)
-    file_name = sys.argv[1]
-    file_size = int(sys.argv[2])
-    
+    file_name = os.getenv('FILE_NAME')
+    file_size = int(os.getenv('BODY_SIZE'))
+    if not file_name or not file_size:
+        print("400")
+        sys.exit(-1)
     data = b""
     bytes_read = 0
     while True:
